Rename hal/ to devices/
