+++ Boards/LilygoTLoraPager/Source/devices/TpagerEncoder.cpp
